refactor: extract ScopeEntry into Scope/ScopeEntry.swift
Part of phase 1 (#756) — RunnerBarCore prep.
Pure code move, no behaviour change.
